+++ src/onegov.gazette/onegov/gazette/tests/test_views_categories.py
@@ -3,6 +3,7 @@ from onegov.gazette.tests.common import login_editor_1
 from onegov.gazette.tests.common import login_publisher
 from pyquery import PyQuery as pq
 from webtest import TestApp as Client
+from xlrd import open_workbook
 
 
 def test_view_categories(gazette_app):
@@ -121,3 +122,47 @@ def test_view_categories_permissions(gazette_app):
     client.get('/categories', status=403)
     client.get(edit_link, status=403)
     client.get(delete_link, status=403)
+
+
+def test_view_categories_export(gazette_app):
+    client = Client(gazette_app)
+
+    client.get('/categories/export', status=403)
+
+    login_editor_1(client)
+    client.get('/categories/export', status=403)
+
+    login_publisher(client)
+    response = client.get('/categories/export')
+
+    book = open_workbook(file_contents=response.body)
+    assert book.nsheets == 1
+
+    sheet = book.sheets()[0]
+    assert sheet.ncols == 4
+    assert sheet.nrows == 6
+
+    assert sheet.cell(0, 0).value == 'ID'
+    assert sheet.cell(0, 1).value == 'Name'
+    assert sheet.cell(0, 2).value == 'Titel'
+    assert sheet.cell(0, 3).value == 'Aktiv'
+
+    assert sheet.cell(1, 1).value == '13'
+    assert sheet.cell(1, 2).value == 'Commercial Register'
+    assert sheet.cell(1, 3).value == 1
+
+    assert sheet.cell(2, 1).value == '10'
+    assert sheet.cell(2, 2).value == 'Complaints'
+    assert sheet.cell(2, 3).value == 0
+
+    assert sheet.cell(3, 1).value == '11'
+    assert sheet.cell(3, 2).value == 'Education'
+    assert sheet.cell(3, 3).value == 1
+
+    assert sheet.cell(4, 1).value == '14'
+    assert sheet.cell(4, 2).value == 'Elections'
+    assert sheet.cell(4, 3).value == 1
+
+    assert sheet.cell(5, 1).value == '12'
+    assert sheet.cell(5, 2).value == 'Submissions'
+    assert sheet.cell(5, 3).value == 1